requirements.txt
asgiref==3.9.2
Django==5.2.6
django-unfold==0.67.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.9
PyJWT==2.10.1
sqlparse==0.5.3
